- (snmplib/snmp_api.h)   - Changed trap_type & specific_type to long from int for 64 bit architectures.     - suggested by francus@metsny.delphi.com.
git-svn-id: file:///home/hardaker/lib/sf-bkups/net-snmp-convert-svnrepo/trunk/net-snmp@317 06827809-a52a-0410-b366-d66718629ded

+++ snmplib/snmp_api.h
@@ -99,8 +99,8 @@ struct snmp_pdu {
     oid	    *enterprise;/* System OID */
     int	    enterprise_length;
     ipaddr  agent_addr;	/* address of object generating trap */
-    int	    trap_type;	/* trap type */
-    int	    specific_type;  /* specific type */
+    long	    trap_type;	/* trap type */
+    long	    specific_type;  /* specific type */
     u_long  time;	/* Uptime */
 
     struct variable_list *variables;
